[SPHKMBH-93] +: Base64 encode all inline scripts.
Only javascript with a source attribute can have a defer attribute. Because of this we can't defer the inline javascript. This is a problem because all the JS libraries are defered. To solve this problem we encode the content of all inline javascript and place this code inside a source attribute. Now with the source attribute we can also defer the inline javascript.